A ShapeResultView's last safe offset must be after the view's start
It used to be that, in cases where a ShapeResultView's text is Latin
characters without ligatures closed to the passed index, the result of
the ShapeResultView::PreviousSafeToBreakOffset was the length of the
view. This could sometimes be before the start of the view, or it
could be at an unexpected position inside the view. In
https://crbug.com/1304876, this was found because the last safe offset
turned out to be in the middle of a ligature, which caused a visible
difference.

Changing PreviousSafeToBreakOffset to have |run_start| be the maximum
of the part's start index and the view's |StartIndex()| seems to fix
it.

Bug: 1304876
Change-Id: Ied68dd5986b603814fa236e7ab2081db5a817723