Turn off independent inheritance if children have explicitly inherited properties.
If any of our children have explicit inheritance of a non-inherited
property, we cannot mark it as using independently inherited properties
only, as some of the explicit properties might inherit (our diffing
considers inherited properties only). We used to gate this behind
a test of diffs in the non-inherited properties, but that would only
consider that difference on this element, not its children (and the
problem could propagate fro there to its grandchildren).

This seems to have been broken in a refactor in
d7e0cc711958dfb3aa39f0f912db1f154617bd98.

Change-Id: If6b790fea727ad68322cde7c4c2a342322283f68